refactor: move developer knowledge client to pkg/clients/dk
